fix: Match multichip levels to 0CC and NSFPlay
- refactor: Chip mix levels now use `double` type
- fix: Typo in setting VRC7 mix level

+++ Source/APU/APU.h
@@ -209,7 +209,7 @@ public:
 		std::vector<int16_t> DeviceMixOffsets
 	);
 
-	void SetChipLevel(chip_level_t Chip, float LeveldB, bool SurveyMix = false);
+	void SetChipLevel(chip_level_t Chip, double LeveldB, bool SurveyMix = false);
 
 	/// Commit changes if no exception is active.
 	///
@@ -227,7 +227,7 @@ private:
 
 	// Mutations.
 	std::optional<uint8_t> m_ExternalSound;
-	std::optional<float> m_ChipLevels[CHIP_LEVEL_COUNT];		// Chip levels, in linear gain factor scale
+	std::optional<double> m_ChipLevels[CHIP_LEVEL_COUNT];		// Chip levels, in linear gain factor scale
 	std::optional<MixerConfig> m_MixerConfig;
 	std::optional<EmulatorConfig> m_EmulatorConfig;
 };
